mvc controller's method getFIbValue modified from random value to real Fibonacci (factory) value. Some code cleanup

+++ src/main/java/pl/sda/mvc/model/Model.java
@@ -7,13 +7,13 @@ import java.util.List;
  * Created by m.losK on 2017-02-25.
  */
 public class Model {
-    public static List<Integer> values = new ArrayList<>();
+    public static List<Long> values = new ArrayList<>();
 
-    public static List<Integer> getValues() {
+    public static List<Long> getValues() {
         return values;
     }
 
-    public static void addValue(int newValue) {
+    public static void addValue(long newValue) {
         values.add(newValue);
     }
 }
