+++ src/main/java/pl/sda/mvc/view/View.java
@@ -1,6 +1,5 @@
 package pl.sda.mvc.view;
 
-import com.sun.scenario.effect.impl.sw.sse.SSEBlend_SRC_OUTPeer;
 import pl.sda.mvc.controller.Controller;
 
 import java.util.Scanner;
@@ -11,12 +10,12 @@ import java.util.Scanner;
 public class View {
     public static void main(String[] args) {
         Scanner in = new Scanner(System.in);
-        while(true){
+        while (true) {
             System.out.println("Enter a number: ");
             Integer choice = in.nextInt();
             Controller controller = new Controller();
-            if(choice!=null && choice>0){
-                int result = controller.getFibValue(choice);
+            if (choice != null && choice > 0) {
+                long result = controller.getFibValue(choice);
                 System.out.println("FIB = " + result);
             } else {
                 System.out.println("AVG = " + controller.average());
